added icon for this addon, implemented proper toString method for Navaid and Waypoint class so that they now display the data in a human-readable format as dd mm ss. closes #2. closes #4

+++ src/main/java/moe/ofs/addon/navdata/RegionBorder.java
@@ -8,8 +8,8 @@ import lombok.ToString;
 @Setter
 @ToString
 public class RegionBorder {
-    private long north;
-    private long south;
-    private long east;
-    private long west;
+    private double north;
+    private double south;
+    private double east;
+    private double west;
 }
